feat: Initialized the repository with the backend
A `package.json` was added into the repository, this is the file that
tracks our dependencies/modules, and is also where we define the
commands to run our application.

We are using the `yarn` utility, you can install this by running,
`npm install yarn`.
In order to run the backend right now, you can use the following
command,
`yarn run backend`.

Finally, I've also added in a basic application to test out whether the
backend works. The structure will be crafted in the following manner for
now,
```
.
├── backend
│   ├── app.js
│   ├── config
│   ├── controllers
│   ├── data
│   ├── middlewares
│   └── routes
├── frontend
├── LICENSE
├── node_modules
├── package.json
└── README.md
```

We'll be placing the components in it's own appropriate directory.